tests/requirements/requirements_base.txt
coverage<5
python-coveralls>2.5.0
unittest-xml-reporting==1.11.0
django-treebeard>=4.3
argparse
dj-database-url
djangocms-admin-style>=1.5
django-admin-rangefilter
django-sekizai>=0.7
django-classy-tags>=0.7.2
iptools
sphinx==1.8.5
sphinxcontrib-spelling<7.0.0 # restriction for py35 tests
pyflakes==1.1.0
pyenchant
django-app-helper
mock>=2.0.0
django-formtools
factory-boy
freezegun

# Unreleased django-cms 4.0 compatible packages
https://github.com/divio/django-cms/tarball/release/4.0.x#egg=django-cms
https://github.com/divio/djangocms-alias/tarball/master#egg=djangocms-alias
https://github.com/divio/djangocms-text-ckeditor/tarball/support/4.0.x#egg=djangocms-text-ckeditor
https://github.com/divio/djangocms-versioning/tarball/master#egg=djangocms-versioning
https://github.com/django-cms/djangocms-moderation/tarball/release/1.0.x#egg=djangocms-moderation
